+++ CHANGELOG.md
@@ -46,6 +46,7 @@
   - `TrimWhitespace()` now supports tracing
   - `Uniq()` now supports tracing
   - `XargsBasename()` now supports tracing
+  - `XargsCat()` now supports tracing
 * Pipelines now set a context flag in their Pipe
   - this is used to tell sinks where to read from!